chore: migrate from tsup to tsdown
tsup is effectively unmaintained and has a bug with TypeScript 6.0
(egoist/tsup#1388 — unconditionally injects baseUrl which triggers
TS5101). tsdown is the spiritual successor, powered by Rolldown, and
supports TS6 out of the box.

- Replace tsup + esbuild with tsdown (rolldown-based bundler)
- Port react-compiler babel plugin from esbuild onLoad to rolldown
  transform hook
- Remove fixCjsDefaultInteropPlugin (rolldown handles CJS interop
  correctly without the isNodeMode bug)
- Update package.json exports to match tsdown output extensions
  (.mjs/.d.mts for ESM, .cjs/.d.cts for CJS)
- Update api-extractor config for new .d.mts entry point
- Remove esbuild from onlyBuiltDependencies in pnpm-workspace.yaml
