+++ packages/seatmaps/api-extractor.json
@@ -1,6 +1,6 @@
 {
     "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-    "mainEntryPointFilePath": "<projectFolder>/dist/index.d.ts",
+    "mainEntryPointFilePath": "<projectFolder>/dist/index.d.mts",
     "apiReport": {
         "enabled": true,
         "reportFileName": "<unscopedPackageName>.api.md",
