+++ packages/seatmaps/package.json
@@ -11,26 +11,26 @@
     "type": "module",
     "exports": {
         ".": {
-            "types": "./dist/index.d.ts",
-            "import": "./dist/index.js",
+            "types": "./dist/index.d.mts",
+            "import": "./dist/index.mjs",
             "require": "./dist/index.cjs"
         },
         "./package.json": "./package.json"
     },
     "main": "./dist/index.cjs",
-    "module": "./dist/index.js",
-    "types": "./dist/index.d.ts",
+    "module": "./dist/index.mjs",
+    "types": "./dist/index.d.mts",
     "files": [
         "dist"
     ],
     "scripts": {
         "api": "api-extractor run --local",
         "api:check": "api-extractor run",
-        "build": "tsup",
+        "build": "tsdown",
         "build-storybook": "storybook build",
         "ch": "pnpm lint && pnpm typecheck && pnpm test && pnpm format:check && pnpm api:check",
         "clean": "rm -rf dist",
-        "dev": "tsup --watch",
+        "dev": "tsdown --watch",
         "format": "prettier --write .",
         "format:check": "prettier --check .",
         "lint": "eslint src",
@@ -57,7 +57,6 @@
         "@types/react-dom": "^19.2.3",
         "@vitejs/plugin-react": "^6.0.1",
         "babel-plugin-react-compiler": "^1.0.0",
-        "esbuild": "^0.27.4",
         "eslint": "^10.1.0",
         "eslint-plugin-react": "^7.35.0",
         "eslint-plugin-react-hooks": "^7.0.1",
@@ -66,7 +65,7 @@
         "react": "^19.2.4",
         "react-dom": "^19.2.4",
         "storybook": "^10.3.1",
-        "tsup": "^8.0.0",
+        "tsdown": "^0.21.7",
         "typescript": "^5.7.0",
         "typescript-eslint": "^8.57.1",
         "vite": "^8.0.5",
