+++ pnpm-workspace.yaml
@@ -1,5 +1,2 @@
 packages:
   - packages/*
-
-onlyBuiltDependencies:
-  - esbuild
